+++ app/controllers/authentication_controller.rb
@@ -1,10 +1,10 @@
 class AuthenticationController < ApplicationController
-  skip_before_action :authenticate_request
+  skip_before_action :validate_json_web_token
 
   def login
     @user = User.find_by_email(params[:email])
     if @user.authenticate(params[:password])
-      token =  jwt_encode(@user.id, nil, 24.hours.from_now)
+      token =  JsonWebToken::UserToken.encode(@user.id, nil, 24.hours.from_now)
       render json: {token: token}, status: :ok
     else
       render json:{error: 'unauthorized'}, status: :unauthorized
